+++ nibRealUI/Util.lua
@@ -387,7 +387,7 @@ function RealUI.TryInCombat(func, alert, ...)
 
         args[func] = {...}
 
-        if not alertShown then
+        if not alertShown or alert == false then
             RealUI:Notification(RealUI.L["Alert_CombatLockdown"], true, alert or RealUI.L["Alert_WaitCombatLockdown"], nil, [[Interface\DialogFrame\UI-Dialog-Icon-AlertNew]])
             alertShown = true
         end